==> 01-creating-master-files.ipynb <==
# Data Preparation — Creating Master Files
The code imports the separate yearly temperature readings and creates three master dataframes: 
- Historic data (1983-2012)
- Recent data (2013-2022)
- Latest incomplete data (January 1, 2023-)

## Importing libraries

import os
import pandas as pd
import glob

## Importing files and creating master dataframes

### Historic temperatures (1983-2012)

#### Joining up all files into single dataframe

### creates empty list to write into
dfs = []

### reads all files from the folder
files = glob.glob("raw_data/yearly_data/historic/*.csv")

### loop runs through all the files
for file in files:
    ### skips first 10 rows and imports all the files
    df = pd.read_csv(file, skiprows = 10)
    ### appends file to the list
    dfs.append(df)

### joins up the files vertically    
data_hist = pd.concat(dfs, ignore_index = True)

### displays dataframe
data_hist

#### Modifying dataframe

### renames columns
data_hist = data_hist.rename(columns = {"Name": "COUNTY", "Longitude": "LONG",
                                        "Latitude": "LAT", "Elevation (ft)": "ELEV", 
                                        "Date": "DATE", "ppt (inches)": "RAINFALL", 
                                        "tmin (degrees F)": "TMIN", "tmean (degrees F)": "TMEAN", 
                                        "tmax (degrees F)": "TMAX", "tdmean (degrees F)": "TDMEAN",
                                        "vpdmin (hPa)": "VPDMIN", "vpdmax (hPa)": "VPDMAX"})

### converts date to date-time format
data_hist["DATE"] = pd.to_datetime(data_hist["DATE"])

### converts and stores dates in YYYY and MM-YYYY formats
data_hist["MONTH_YEAR"] = pd.to_datetime(data_hist["DATE"]).dt.to_period("M")
data_hist["YEAR"] = pd.to_datetime(data_hist["DATE"]).dt.to_period("Y")

### converts and stores the dates as strings
data_hist["MONTH_YEAR_STR"] = data_hist["MONTH_YEAR"].astype(str)
data_hist["YEAR_STR"] = data_hist["YEAR"].astype(str)

### displays dataframe
data_hist

### Recent temperatures (2013-2022)

#### Joining up all files into single dataframe

### creates empty list to write into
dfs = []

### reads all files from the folder
files = glob.glob("raw_data/yearly_data/recent/*.csv")

### loop runs through all the files
for file in files:
    ### skips first 10 rows and imports all the files
    df = pd.read_csv(file, skiprows = 10)
    ### appends file to the list
    dfs.append(df)

### joins up the files vertically   
data_rec = pd.concat(dfs, ignore_index = True)

### displays dataframe
data_rec

#### Modifying dataframe

### renames columns
data_rec = data_rec.rename(columns = {"Name": "COUNTY", "Longitude": "LONG",
                                        "Latitude": "LAT", "Elevation (ft)": "ELEV", 
                                        "Date": "DATE", "ppt (inches)": "RAINFALL", 
                                        "tmin (degrees F)": "TMIN", "tmean (degrees F)": "TMEAN", 
                                        "tmax (degrees F)": "TMAX", "tdmean (degrees F)": "TDMEAN",
                                        "vpdmin (hPa)": "VPDMIN", "vpdmax (hPa)": "VPDMAX"})
### converts date to date-time format
data_rec["DATE"] = pd.to_datetime(data_rec["DATE"])

### converts and stores dates in YYYY and MM-YYYY formats
data_rec["MONTH_YEAR"] = pd.to_datetime(data_rec["DATE"]).dt.to_period("M")
data_rec["YEAR"] = pd.to_datetime(data_rec["DATE"]).dt.to_period("Y")

### converts and stores the dates as strings
data_rec["MONTH_YEAR_STR"] = data_rec["MONTH_YEAR"].astype(str)
data_rec["YEAR_STR"] = data_rec["YEAR"].astype(str)

### displays dataframe
data_rec

### Latest temperatures (January 1, 2023-)

#### Importing data

data_lat = pd.read_csv("raw_data/yearly_data/2023_latest/2023.csv", skiprows = 10)

data_lat

#### Modifying dataframe

### renames columns
data_lat = data_lat.rename(columns = {"Name": "COUNTY", "Longitude": "LONG",
                                        "Latitude": "LAT", "Elevation (ft)": "ELEV", 
                                        "Date": "DATE", "ppt (inches)": "RAINFALL", 
                                        "tmin (degrees F)": "TMIN", "tmean (degrees F)": "TMEAN", 
                                        "tmax (degrees F)": "TMAX", "tdmean (degrees F)": "TDMEAN",
                                        "vpdmin (hPa)": "VPDMIN", "vpdmax (hPa)": "VPDMAX"})
### converts date to date-time format
data_lat["DATE"] = pd.to_datetime(data_lat["DATE"])

### converts and stores dates in YYYY and MM-YYYY formats
data_lat["MONTH_YEAR"] = pd.to_datetime(data_lat["DATE"]).dt.to_period("M")
data_lat["YEAR"] = pd.to_datetime(data_lat["DATE"]).dt.to_period("Y")

### converts and stores the dates as strings
data_lat["MONTH_YEAR_STR"] = data_lat["MONTH_YEAR"].astype(str)
data_lat["YEAR_STR"] = data_lat["YEAR"].astype(str)

### displays dataframe
data_lat

## Exporting master dataframes

data_hist.to_csv("raw_data/master_hist.csv", index = False)
data_rec.to_csv("raw_data/master_rec.csv", index = False)
data_lat.to_csv("raw_data/master_lat.csv", index = False)
