==> 02-computing-heat-indices.ipynb <==
# Computing Heat Indices

The code calculates a measure — the heat index — for what the temperature "feels like" to the human body when relative humidity is combined with the air temperature.

## Importing libraries

import pandas as pd
import time
from modules import relHumidity, heatIndex

## Importing dataframes

### imports historical data
data_hist = pd.read_csv("raw_data/master_hist.csv")

### displays dataframe
data_hist

### imports recent data
data_rec = pd.read_csv("raw_data/master_rec.csv")

### displays dataframe
data_rec

### imports latest data
data_lat = pd.read_csv("raw_data/master_lat.csv")

### displays dataframe
data_lat

## Calculating heat indices

### Storing lengths of dataframe

### stores length of historical dataframe
historic_len = len(data_hist)

### stores length of recent dataframe
recent_len = len(data_rec)

### stores length of latest dataframe
lat_len = len(data_lat)

### Defining loops to compute heat indices and relative humidity¶

#### Historical temperatures (1983-2012)

### creates columns to write into
data_hist["TMIN_INDEX"] = ""
data_hist["TMAX_INDEX"] = ""
data_hist["TMEAN_INDEX"] = ""
data_hist["REL_HUMIDITY"] = ""

### loop runs through dataframe
for i in range(0, historic_len):
    
    ### stores mean temperature and mean dewpoint
    temp_mean = data_hist["TMEAN"][i]
    dew = data_hist["TDMEAN"][i]
    
    ### calls predefined method and computes relative humidity
    data_hist["REL_HUMIDITY"][i] = round(relHumidity(temp_mean, dew), 1)
    
    ### calls predefined method and computes heat indices 
    data_hist["TMIN_INDEX"][i] = heatIndex(data_hist["TMIN"][i], data_hist["REL_HUMIDITY"][i])
    data_hist["TMAX_INDEX"][i] = heatIndex(data_hist["TMAX"][i], data_hist["REL_HUMIDITY"][i])
    data_hist["TMEAN_INDEX"][i] = heatIndex(data_hist["TMEAN"][i], data_hist["REL_HUMIDITY"][i])
    
    ### prints out a progress report showing number of rows processed    
    if(i%10000 == 0):
        print("Completed: ", str(int(i/historic_len * 100)), "%")
        print("Rows completed: ", str(i))
        print("_______")
    elif(i == (historic_len - 1)):
        print("Completed: 100%")
        print("_______")
        print("_______") 

#### Recent temperatures (2013-2022)

### creates columns to write into
data_rec["TMIN_INDEX"] = " "
data_rec["TMAX_INDEX"] = " "
data_rec["TMEAN_INDEX"] = " "
data_rec["REL_HUMIDITY"] = " "

### loop runs through dataframe
for i in range(0, recent_len):
    
    ### stores mean temperature and mean dewpoint
    temp_mean = data_rec["TMEAN"][i]
    dew = data_rec["TDMEAN"][i]
    
    ### calls predefined method and computes relative humidity
    data_rec["REL_HUMIDITY"][i] = round(relHumidity(temp_mean, dew), 1)
    
    ### calls predefined method and computes heat indices 
    data_rec["TMIN_INDEX"][i] = heatIndex(data_rec["TMIN"][i], data_rec["REL_HUMIDITY"][i])
    data_rec["TMAX_INDEX"][i] = heatIndex(data_rec["TMAX"][i], data_rec["REL_HUMIDITY"][i])
    data_rec["TMEAN_INDEX"][i] = heatIndex(data_rec["TMEAN"][i], data_rec["REL_HUMIDITY"][i])
    
    ### prints out a progress report showing number of rows processed    
    if(i%10000 == 0):
        print("Completed: ", str(int(i/recent_len * 100)), "%")
        print("Rows completed: ", str(i))
        print("_______")
    elif(i == (recent_len - 1)):
        print("Completed: 100%")
        print("_______")
        print("_______") 

#### Latest temperatures (January 1, 2023 — June 14, 2023)

### creates columns to write into
data_lat["TMIN_INDEX"] = " "
data_lat["TMAX_INDEX"] = " "
data_lat["TMEAN_INDEX"] = " "
data_lat["REL_HUMIDITY"] = " "

### loop runs through dataframe
for i in range(0, lat_len):
    
    ### stores mean temperature and mean dewpoint
    temp_mean = data_lat["TMEAN"][i]
    dew = data_lat["TDMEAN"][i]
    
    ### calls predefined method and computes relative humidity
    data_lat["REL_HUMIDITY"][i] = round(relHumidity(temp_mean, dew), 1)

    ### calls predefined method and computes heat indices
    data_lat["TMIN_INDEX"][i] = heatIndex(data_lat["TMIN"][i], data_lat["REL_HUMIDITY"][i])
    data_lat["TMAX_INDEX"][i] = heatIndex(data_lat["TMAX"][i], data_lat["REL_HUMIDITY"][i])
    data_lat["TMEAN_INDEX"][i] = heatIndex(data_lat["TMEAN"][i], data_lat["REL_HUMIDITY"][i])
    
    ### prints out a progress report showing number of rows processed    
    if(i%1000 == 0):
        print("Completed: ", str(int(i/lat_len * 100)), "%")
        print("Rows completed: ", str(i))
        print("_______")
    elif(i == (lat_len - 1)):
        print("Completed: 100%")
        print("_______")
        print("_______") 

## Displaying final dataframes

data_hist

data_rec

data_lat

## Exporting dataframes

data_hist.to_csv("generated_data/historic_hi.csv", index = False)
data_rec.to_csv("generated_data/recent_hi.csv", index = False)
data_lat.to_csv("generated_data/latest_hi.csv", index = False)
