==> 03-state-averages-analysis.ipynb <==
# State Average Temperatures — Analysis

The code analyzes the daily temperature data to compute Florida's daily average temperatures and heat indices. 

Read more on the thresholds [here](https://web.archive.org/web/20230619070053/https://www.noaa.gov/jetstream/global/heat-index).

## Importing libraries

import pandas as pd
import time
from modules import relHumidity, heatIndex

## Importing datasets and creating master dataframe

data_hist = pd.read_csv("generated_data/historic_hi.csv")
data_rec = pd.read_csv("generated_data/recent_hi.csv")

Note: Data from 2023 is excluded since it is incomplete

### creating master dataframe
master = pd.concat([data_hist, data_rec], ignore_index = True)

### displays dataframe
master

## Computing average daily state figures

### Creates dataframe

### selects columns
columns = ["TMIN", "TMAX", "TMEAN", "TDMEAN"]

### computes mean values for that date
state_daily_avg = master.groupby("DATE", as_index = False)[columns].mean()

### displays dataframe
state_daily_avg

### Computing heat indices

### computes length of dataframe
len_st_daily_avg = len(state_daily_avg)

### creates columns to write into
state_daily_avg["TMIN_INDEX"] = ""
state_daily_avg["TMAX_INDEX"] = ""
state_daily_avg["TMEAN_INDEX"] = ""
state_daily_avg["REL_HUMIDITY"] = ""

### loop runs through dataframe
for i in range(0, len_st_daily_avg):
    
    ### stores mean temperature and mean dewpoint
    temp_mean = state_daily_avg["TMEAN"][i]
    dew = state_daily_avg["TDMEAN"][i]
    
    ### calls predefined method and computes relative humidity
    state_daily_avg["REL_HUMIDITY"][i] = round(relHumidity(temp_mean, dew), 1)
    
    ### calls predefined method and computes heat indices 
    state_daily_avg["TMIN_INDEX"][i] = heatIndex(state_daily_avg["TMIN"][i], state_daily_avg["REL_HUMIDITY"][i])
    state_daily_avg["TMAX_INDEX"][i] = heatIndex(state_daily_avg["TMAX"][i], state_daily_avg["REL_HUMIDITY"][i])
    state_daily_avg["TMEAN_INDEX"][i] = heatIndex(state_daily_avg["TMEAN"][i], state_daily_avg["REL_HUMIDITY"][i])
    
    ### prints out a progress report showing number of rows processed    
    if(i%500 == 0):
        print("Completed: ", str(int(i/len_st_daily_avg * 100)), "%")
        print("Rows completed: ", str(i))
        print("_______")
    elif(i == (len_st_daily_avg - 1)):
        print("Completed: 100%")
        print("_______")
        print("_______") 

### displays dataframe
state_daily_avg

### Finalizing dataframe

### converts date to date-time format
state_daily_avg["DATE"] = pd.to_datetime(state_daily_avg["DATE"])

### converts and stores dates in YYYY and MM-YYYY formats
state_daily_avg["MONTH_YEAR"] = pd.to_datetime(state_daily_avg["DATE"]).dt.to_period("M")
state_daily_avg["YEAR"] = pd.to_datetime(state_daily_avg["DATE"]).dt.to_period("Y")

### converts and stores the dates as strings
state_daily_avg["MONTH_YEAR_STR"] = state_daily_avg["MONTH_YEAR"].astype(str)
state_daily_avg["YEAR_STR"] = state_daily_avg["YEAR"].astype(str)

### displays dataframe
state_daily_avg

### Exporting dataframe

state_daily_avg.to_csv("generated_data/state_daily_avg.csv", index = False)

## Computing overall state figures for 1983 to 2022

### Average temperature and dew point temperature

### computes average temperatures
avg_temp = round(state_daily_avg["TMEAN"].mean(), 3)
avg_dp = round(state_daily_avg["TDMEAN"].mean(), 3)

print("Average temperature (1983-2022):", avg_temp)
print("Average dew point temperature (1983-2022):", avg_dp)

### Average relative humidity

### computes relative humidity 
rel_humidity = round(relHumidity(avg_temp, avg_dp), 1)

print("Average relative humidity (1983-2022):", rel_humidity)

### Average heat index

### computes heat index
hi_avg = heatIndex(avg_temp, rel_humidity)

print("Average heat index (1983-2022):", hi_avg)

## Computing average yearly figures

### Creating dataframe

### displays dataframe
state_daily_avg.head()

### selects columns 
columns = ["TMEAN", "TDMEAN"]

### computes yearly averages
state_yearly_avg = state_daily_avg.groupby("YEAR_STR", as_index = False)[columns].mean()

### displays dataframe
state_yearly_avg

### Computing heat indices

### creates column to write into
state_yearly_avg["TMEAN_INDEX"] = " "

### loop runs through dataframe
for i in range(0, len(state_yearly_avg)):
    
    ### stores mean temperature and mean dewpoint
    temp_mean = state_yearly_avg["TMEAN"][i]
    dew = state_yearly_avg["TDMEAN"][i]
    
    ### calls predefined method and computes relative humidity
    rh = round(relHumidity(temp_mean, dew), 1)
    
    ### calls predefined method and computes heat indices
    state_yearly_avg["TMEAN_INDEX"][i] = heatIndex(state_yearly_avg["TMEAN"][i], rh)

state_yearly_avg.head()

### Computing deviations

### creates column to write into
state_yearly_avg["OVERALL_AVG_TEMP"] = avg_temp
state_yearly_avg["OVERALL_AVG_INDEX"] = hi_avg
state_yearly_avg["DEVIATION_TEMP"] = ""
state_yearly_avg["DEVIATION_INDEX"] = ""

### loop runs through dataframe
for i in range(0, len(state_yearly_avg)):
    
    ### computes and stores difference in temperature between that year and the overall state's average
    state_yearly_avg["DEVIATION_TEMP"][i] = state_yearly_avg["TMEAN"][i] - avg_temp
    
    ### computes and stores difference in heat index between that year and the overall state's average
    state_yearly_avg["DEVIATION_INDEX"][i] = state_yearly_avg["TMEAN_INDEX"][i] - hi_avg

### displays dataframe
state_yearly_avg

## Exporting dataframes

state_daily_avg.to_csv("generated_data/state_daily_avg.csv", index = False)
state_yearly_avg.to_csv("generated_data/state_yearly_avg.csv", index = False)
