==> 06-dade-medians-analysis.ipynb <==
# Miami-Dade Median Temperatures — Analysis

The code analyzes the daily temperature data to compute Miami-Dade county's daily median temperatures, heat indices and days above certain threshold heat index values. 

Read more on the thresholds [here](https://web.archive.org/web/20230619070053/https://www.noaa.gov/jetstream/global/heat-index).

## Importing libraries

import pandas as pd
import time
from modules import relHumidity, heatIndex

## Importing datasets and creating master dataframe

data_hist = pd.read_csv("generated_data/historic_hi.csv")
data_rec = pd.read_csv("generated_data/recent_hi.csv")

Note: Data from 2023 is excluded since it is incomplete

### creating master dataframe
master = pd.concat([data_hist, data_rec], ignore_index = True)

### filtering for county
dade_daily_med = master[master["COUNTY"] == "Miami-Dade"].reset_index(drop = True)

### displayes dataframe
dade_daily_med

## Computing overall county figures for 1983 to 2022

### Median temperature and dewpoint temperature

### computes average temperatures
med_temp = round(dade_daily_med["TMEAN"].median(), 3)
med_dp = round(dade_daily_med["TDMEAN"].median(), 3)

print("Median temperature (1983-2022):", med_temp)
print("Median dewpoint temperature (1983-2022):", med_dp)

### Average relative humidity

### computes relative humidity 
rel_humidity = round(relHumidity(med_temp, med_dp), 1)

print("Median relative humidity (1983-2022):", rel_humidity)

### Average heat index

### computes heat index
hi_med = heatIndex(med_temp, rel_humidity)

print("Median heat index (1983-2022):", hi_med)

## Computing average yearly figures

### Creating dataframe

### selects columns 
columns = ["TMEAN", "TDMEAN"]

### computes yearly averages
dade_yearly_med = dade_daily_med.groupby("YEAR_STR", as_index = False)[columns].median()

### displays dataframe
dade_yearly_med

### Computing heat indices

### creates column to write into
dade_yearly_med["TMEAN_INDEX"] = " "

### loop runs through dataframe
for i in range(0, len(dade_yearly_med)):
    
    ### stores mean temperature and median dewpoint
    temp_med = dade_yearly_med["TMEAN"][i]
    dew = dade_yearly_med["TDMEAN"][i]
    
    rh = round(relHumidity(temp_med, dew), 1)
    
    ### calls method to compute heat indices and stores the values
    dade_yearly_med["TMEAN_INDEX"][i] = heatIndex(dade_yearly_med["TMEAN"][i], rh)

dade_yearly_med.head()

### Computing deviations

### creates column to write into
dade_yearly_med["OVERALL_MED_TEMP"] = med_temp
dade_yearly_med["OVERALL_MED_INDEX"] = hi_med
dade_yearly_med["DEVIATION_TEMP"] = ""
dade_yearly_med["DEVIATION_INDEX"] = ""

### loop runs through dataframe
for i in range(0, len(dade_yearly_med)):
    
    ### computes and stores difference in temperature between that year and the overall county's average
    dade_yearly_med["DEVIATION_TEMP"][i] = dade_yearly_med["TMEAN"][i] - med_temp
    
    ### computes and stores difference in heat index between that year and the overall county's average
    dade_yearly_med["DEVIATION_INDEX"][i] = dade_yearly_med["TMEAN_INDEX"][i] - hi_med

### displays dataframe
dade_yearly_med

## Computing day counts above certain thresholds

### displays daily temperature dataframe
dade_daily_med.head()

### Mean heat index >= 90°F

#### Filtering data

### filters dataframe
mean90up = dade_daily_med[dade_daily_med["TMEAN_INDEX"] >= 90].reset_index(drop = True)

### displays dataframe
mean90up

#### Creating dataframe

### computes yearly figures
days_90up = mean90up["YEAR_STR"].value_counts().sort_index(ascending = True).to_frame().reset_index()

### renames columns
days_90up.columns = ["YEAR", "MEANINDEX_90UP"]

### displays dataframe
days_90up.head()

### Minimum heat index >= 105°F

### filters dataframe
dade_daily_med[dade_daily_med["TMIN_INDEX"] >= 105].reset_index(drop = True)

There were no days which meet the criterion.

### Mean heat index >= 105°F

### filters dataframe
dade_daily_med[dade_daily_med["TMEAN_INDEX"] >= 105].reset_index(drop = True)

There were no days which meet the criterion.

### Maximum heat index >= 105°F

#### Filtering data

### filters dataframe
days105up = dade_daily_med[dade_daily_med["TMAX_INDEX"] >= 105].reset_index(drop = True)

### displays dataframe
days105up

#### Creating dataframe

### computes yearly figures
days_105up = days105up["YEAR_STR"].value_counts().sort_index(ascending = True).to_frame().reset_index()

### renames columns
days_105up.columns = ["YEAR", "MAXINDEX_105UP"]

### displays dataframe
days_105up.head()

### Maximum heat index >= 130°F

#### Filtering data

### filters dataframe
days130up = dade_daily_med[dade_daily_med["TMAX_INDEX"] >= 130].reset_index(drop = True)

### displays dataframe
days130up

#### Creating dataframe

### computes yearly figures
days_130up = days130up["YEAR_STR"].value_counts().sort_index(ascending = True).to_frame().reset_index()

### renames columns
days_130up.columns = ["YEAR", "MAXINDEX_130UP"]

### displays dataframe
days_130up.head()

### Merging dataframes

### merges dataframes
yearly_dayscount = pd.merge(days_90up, days_105up, on = "YEAR", how = "outer")
yearly_dayscount = pd.merge(yearly_dayscount, days_130up, on = "YEAR", how = "outer")

### resets index
yearly_dayscount = yearly_dayscount.sort_values("YEAR").reset_index(drop = True)

### adds columns for filters with no results
yearly_dayscount["MININDEX_105UP"] = 0
yearly_dayscount["MEANINDEX_105UP"] = 0

### controls for null values
yearly_dayscount = yearly_dayscount.fillna(0)
yearly_dayscount = yearly_dayscount.round(0).astype(int)

leap = [1984, 1988, 1992, 1996, 2000, 2004, 2008, 2012, 2016, 2020]
yearly_dayscount["TOTAL_DAYS"] = 365
for i in range(0, len(yearly_dayscount)):
    
    if yearly_dayscount["YEAR"][i] in leap:
        yearly_dayscount["TOTAL_DAYS"][i] = 366

### displays dataframe
yearly_dayscount

## Exporting dataframes

dade_daily_med.to_csv("generated_data/dade_daily_med.csv", index = False)
dade_yearly_med.to_csv("generated_data/dade_yearly_med.csv", index = False)
yearly_dayscount.to_csv("generated_data/dade_dayscount.csv", index = False)
