==> 07-consecutive-hot-days-analysis.ipynb <==
# Consecutive Hot Days, 2013-2022 — Analysis

The code computes the number of consecutive days where the heat index was above 105°F — the limit above which sunstrokes, heat cramps, or heat exhaustion is likely and heat stroke is possible with prolonged exposure and/or physical activity. It also computes the average number of instances of these consecutive days per year for each of the counties and how high the temperatures went during that period.

Read more on the thresholds [here](https://web.archive.org/web/20230619070053/https://www.noaa.gov/jetstream/global/heat-index).

## Importing libraries

import pandas as pd
import time
import statistics
from statistics import mode
import numpy as np

## Importing dataset

data_rec = pd.read_csv("generated_data/recent_hi.csv")
data_rec = data_rec.sort_values(by = ["COUNTY", "DATE"], ascending = True).reset_index(drop = True) 

data_rec

## Marking consecutive days with heat index >= 105°F

### Creating copy of dataframe

df_recent = data_rec.copy()

### Creating list of counties

### loop runs through dataframe
for col in df_recent:
    ### stores county name
    counties = df_recent["COUNTY"].unique()

### prints recorded data
print("Counties: " + str(len(counties)))
print(counties)

### Recording consecutive days

### creates column to mark consecutive days with default as "N"
df_recent["CONSECUTIVE"] = "N"

### loop runs through each row in the dataframe containing county names
for i in range(0, len(counties)):
    
    ### stores county 
    location = counties[i]
    
    ### loop runs through each row in the temperature dataframe
    for j in range(0, (len(df_recent)-1)):
        
        ### checks if the coordinates of two consecutive rows are the same as the location stored from the threshold value dataframe row
        if (df_recent["COUNTY"][j] == location) and (df_recent["COUNTY"][j+1] == location):
            
            ### checks if the maximum heat index of two consecutive rows are above the threshold value
            if df_recent["TMAX_INDEX"][j] >= 105 and df_recent["TMAX_INDEX"][j+1] >= 105:
                
                ### marks both rows as consecutive
                df_recent["CONSECUTIVE"][j] = "Y"
                df_recent["CONSECUTIVE"][j+1] = "Y"

### displays dataframe
df_recent[df_recent["CONSECUTIVE"] == "Y"]

### Filtering for only consecutive days above threshold

### creates new dataframe after dropping all rows which are not above threshold
df_recent_H = df_recent[(df_recent["CONSECUTIVE"] == "Y")]
df_recent_H.reset_index(drop = True, inplace = True) 

### displays dataframe
df_recent_H

## Computing number of instances of consecutive days above threshold per county

### creates new column to be written into
df_recent_H["SL_COUNTER"] = ""

### converts data in the "DATE" column to date-time format
df_recent_H["DATE"] =  pd.to_datetime(df_recent_H["DATE"], infer_datetime_format = True)

### Recording number of instances

### loop runs through each row in the dataset containing county names
for i in range(0, len(counties)):
    
    ### stores county
    location = counties[i]
    
    ### initializes a variable to count number of heat waves in that location
    count = 1
    
    ### loop runs through each row in the temperature dataframe
    for j in range(0, (len(df_recent_H)-1)):

        ### checks if the coordinates of two consecutive rows are the same as the location stored from the threshold value dataframe row
        if (df_recent_H["COUNTY"][j] == location) and (df_recent_H["COUNTY"][j+1] == location):
            
                ### calculates and stores the difference between the two dates in the consecutive row
                result = str(df_recent_H["DATE"][j+1] - df_recent_H["DATE"][j])
                
                ### checks if the difference between the dates is 1 day
                if result == "1 days 00:00:00":
                    
                    ### marks both as experiencing the same heat wave
                    df_recent_H["SL_COUNTER"][j] = count
                    df_recent_H["SL_COUNTER"][j+1] = count
                
                ### moves on to next heat wave number if condition is not met
                else:
                    ### increases counter
                    count = count + 1

### displays dataframe
df_recent_H

### Exporting dataframe

df_recent_H.to_csv("generated_data/hotdays_consecutive.csv", index = False)

## Computing county figures

### Creating spreadsheet to write into

with open("generated_data/hotdays_countydetails.csv", "w") as f1:
    f1.write("COUNTY" + "\t" + "MIN_TEMP_HIGHEST" + "\t" + "MAX_TEMP_HIGHEST" + "\t" + "MEAN_TEMP_HIGHEST" + "\t" + "MIN_INDEX_HIGHEST" + "\t" + "MAX_INDEX_HIGHEST" + "\t" + "MEAN_INDEX_HIGHEST" + "\t" + "SL_COUNTER" + "\t" + "HOT_AVG" + "\t" + "HOT_LONGEST" + "\t" + "HOT_LONGEST_DAYS" + "\n")

### Computing figures and recording in spreadsheet

### loop runs through each row in the dataset containing county names
for i in range(0, len(counties)):
    
    ### stores county
    location = counties[i]
    
    ### creates empty list to later store temperatures
    temps_min = []
    temps_max = []
    temps_mean = []
    
    ### creates empty list to later store heat indices
    index_min = []
    index_max = []
    index_mean = []

    ### creates empty list to later store the serial number from the heat wave counter
    temps_hot_no = []
    
    ### loop runs through each row of the above-threshold dataframe
    for j in range(0, len(df_recent_H)):
        
        ### checks if the coordinates is the same as the stored location 
        if df_recent_H["COUNTY"][j] == location:
            
            ### fills up temperature details in the created lists
            temps_min.append(df_recent_H["TMIN"][j])
            temps_max.append(df_recent_H["TMAX"][j])
            temps_mean.append(df_recent_H["TMEAN"][j])
            
            ### fills up heat indices details in the created lists
            index_min.append(df_recent_H["TMIN_INDEX"][j])
            index_max.append(df_recent_H["TMAX_INDEX"][j])
            index_mean.append(df_recent_H["TMEAN_INDEX"][j])

            ### fills up instance number in the created list
            temps_hot_no.append(df_recent_H["SL_COUNTER"][j])
    
    ### calculates and stores the highest temperatures, heat indices and threshold values recorded during the periods
    min_max = np.max(temps_min)
    max_max = np.max(temps_max)
    mean_max = np.max(temps_mean)
    indexmin_max = np.max(index_min)
    indexmax_max = np.max(index_max)
    indexmean_max = np.max(index_mean)
    
    ### stores the total number of instances of consecutive hot days per location
    hot_count = np.max(temps_hot_no)
    ### calculates and stores the average number of instances per year
    hot_avg = int(hot_count/10)
    
    ### stores the instance number for the one that lasts the longest
    hot_mode = mode(temps_hot_no)
    ### calculates the number of days the longest instance lasted
    hot_max = temps_hot_no.count(hot_mode)
    
    ### writes into spreadsheet
    with open("generated_data/hotdays_countydetails.csv", "a") as f2:
        f2.write(str(location) + "\t" + str(min_max) + "\t" + str(max_max) + "\t" + str(mean_max) + "\t"
                 + str(indexmin_max) + "\t" + str(indexmax_max) + "\t" + str(indexmean_max) + "\t" + str(hot_count) + 
                 "\t" + str(hot_avg) + "\t" + str(hot_mode) + "\t" + str(hot_max) + "\n")

### Imports and displays dataset

df_hotdays_details = pd.read_csv("generated_data/hotdays_countydetails.csv", sep = "\t")

df_hotdays_details
